roffit: Change pre tag indent formatting method
- Add indent class for pre tags (piggyback on p tag indent class).
- Reference the indent class in pre tags.
- Stop using a nested p tag to indent in a pre tag.

Prior to this change the roffit script's output for the pre tag used a
nested p tag which included the css style indent level the text in the
pre tag should appear at. W3C's validator says that isn't correct, even
though it appears to work fine. In order to be correct in this it seems
necessary to reference the indent class in the pre tag directly and do
away with the nested p tag. The roffit output when rendered appears to
be visually the same.